Enable creating and editing of kubernetes connections
