+++ backend/src/main/kotlin/dev/kviklet/kviklet/controller/ExecutionRequestController.kt
@@ -48,14 +48,14 @@ sealed class CreateExecutionRequestRequest(
     open val connectionId: ConnectionId,
     open val title: String,
     open val type: RequestType,
-    open val description: String?,
+    open val description: String,
 )
 
 data class CreateDatasourceExecutionRequestRequest(
     override val connectionId: ConnectionId,
     override val title: String,
     override val type: RequestType,
-    override val description: String?,
+    override val description: String,
     val statement: String?,
     val readOnly: Boolean,
 ) : CreateExecutionRequestRequest(connectionId, title, type, description)
@@ -64,7 +64,7 @@ data class CreateKubernetesExecutionRequestRequest(
     override val connectionId: ConnectionId,
     override val title: String,
     override val type: RequestType,
-    override val description: String?,
+    override val description: String,
     val namespace: String?,
     val podName: String?,
     val containerName: String?,
