+++ backend/src/main/kotlin/dev/kviklet/kviklet/service/IdResolver.kt
@@ -12,14 +12,14 @@ import org.springframework.stereotype.Service
 
 @Service
 class IdResolver(
-    private val datasourceConnectionAdapter: ConnectionAdapter,
+    private val connectionAdapter: ConnectionAdapter,
     private val executionRequestAdapter: ExecutionRequestAdapter,
     private val userAdapter: UserAdapter,
 ) {
     fun resolve(id: SecuredDomainId): SecuredDomainObject? {
         return try {
             when (id) {
-                is ConnectionId -> datasourceConnectionAdapter.getConnection(id)
+                is ConnectionId -> connectionAdapter.getConnection(id)
                 is ExecutionRequestId -> executionRequestAdapter.getExecutionRequestDetails(id)
                 is UserId -> userAdapter.findById(id.toString())
                 else -> throw IllegalArgumentException("Unknown id type: ${id::class}")
